TFRecModel/src/com/sparrowrecsys/offline/tensorflow/my_model.py
import tensorflow as tf


def create_model():
    model = tf.keras.models.Sequential(
        [
            tf.keras.layers.Flatten(input_shape=(28, 28)),
            tf.keras.layers.Dense(128, activation="relu"),
            tf.keras.layers.Dropout(0.2),
            tf.keras.layers.Dense(10, activation="softmax", name="output"),
        ]
    )
    # predictions = model(tf.ones((1, 28, 28)))

    model.compile(
        optimizer="adam", loss="sparse_categorical_crossentropy", metrics=["accuracy"]
    )

    return model
